# combobox: Improve performance of option rendering, filtering and string matching