--- a/.changeset/thick-avocados-return.md
+++ b/.changeset/thick-avocados-return.md
@@ -1,0 +1,5 @@
+---
+'@ag.ds-next/react': minor
+---
+
+combobox: Improve performance of option rendering, filtering and string matching.