--- a/packages/react/src/combobox/ComboboxRenderItem.tsx
+++ b/packages/react/src/combobox/ComboboxRenderItem.tsx
@@ -1,13 +1,13 @@
 import { Fragment, isValidElement, ReactNode } from 'react';
-import { Text } from '../text';
-import { Stack } from '../stack';
-import { splitLabel } from './utils';
+import { boxPalette, tokens } from '../core';
+import { fontGrid } from '../core/utils/fontGrid';
 
 export type ComboboxRenderItemProps = {
 	/** The label of the item. */
 	itemLabel: string;
+	// TODO: Deprecate
 	/** The value of the Combobox/Autocomplete text input. */
-	inputValue: string;
+	inputValue?: string;
 	/** Supporting text for the item. */
 	tertiaryText?: ReactNode;
 	/** Supporting text for the item. */
@@ -18,9 +18,10 @@ export type ComboboxRenderItemProps = {
 	endElement?: ReactNode;
 };
 
+const FONT_SIZE_LINE_HEIGHT = fontGrid('sm', 'default');
+
 export function ComboboxRenderItem({
 	itemLabel,
-	inputValue,
 	beforeElement,
 	endElement,
 	secondaryText,
@@ -31,54 +32,69 @@ export function ComboboxRenderItem({
 			{beforeElement ? (
 				<div css={{ flexShrink: 0 }}>{beforeElement}</div>
 			) : null}
-			<Stack as="span">
-				<span>{renderItemLabel(itemLabel, inputValue)}</span>
+			<span
+				css={{
+					alignItems: 'stretch',
+					display: 'flex',
+					flexDirection: 'column',
+				}}
+			>
+				<span>{renderItemLabel(itemLabel)}</span>
 				{secondaryText ? (
 					isValidElement(secondaryText) ? (
 						secondaryText
 					) : (
-						<Text color="muted" fontSize="xs">
+						<span
+							css={{
+								color: boxPalette.foregroundMuted,
+								fontFamily: tokens.font.body,
+								fontWeight: tokens.fontWeight.normal,
+								...FONT_SIZE_LINE_HEIGHT,
+							}}
+						>
 							{secondaryText}
-						</Text>
+						</span>
 					)
 				) : null}
 				{tertiaryText ? (
 					isValidElement(tertiaryText) ? (
 						tertiaryText
 					) : (
-						<Text color="muted" fontSize="xs">
+						<span
+							css={{
+								color: boxPalette.foregroundMuted,
+								fontFamily: tokens.font.body,
+								fontWeight: tokens.fontWeight.normal,
+								...FONT_SIZE_LINE_HEIGHT,
+							}}
+						>
 							{tertiaryText}
-						</Text>
+						</span>
 					)
 				) : null}
-			</Stack>
+			</span>
 			{endElement ? (
-				<div css={{ marginLeft: 'auto', flexShrink: 0 }}>{endElement}</div>
+				<div css={{ flexShrink: 0, marginLeft: 'auto' }}>{endElement}</div>
 			) : null}
 		</Fragment>
 	);
 }
 
-function renderItemLabel(itemLabel: string, inputValue: string) {
-	return splitLabel(itemLabel, inputValue).map((part, index) => {
-		const isHighlighted = part.toLowerCase() === inputValue.toLowerCase();
-		if (isHighlighted) {
-			return (
-				<Text
-					key={index}
-					as="mark"
-					color="inherit"
-					fontWeight="bold"
-					css={{ background: 'none' }}
-				>
-					{part}
-				</Text>
-			);
-		}
-		return (
-			<Text key={index} as="span" color="inherit">
-				{part}
-			</Text>
-		);
-	});
+function renderItemLabel(itemLabel: string) {
+	return (
+		<span
+			css={{
+				color: 'inherit',
+				fontFamily: tokens.font.body,
+				fontWeight: tokens.fontWeight.normal,
+				...FONT_SIZE_LINE_HEIGHT,
+			}}
+		>
+			{itemLabel.split('').map((character, index) => (
+				<span data-char={character.toLowerCase()} key={index}>
+					{character}
+				</span>
+			))}
+		</span>
+	);
 }