--- a/packages/react/src/combobox/utils.ts
+++ b/packages/react/src/combobox/utils.ts
@@ -1,3 +1,4 @@
+import { type CSSProperties } from 'react';
 import { useId } from '../core';
 
 export function useComboboxInputId(idProp?: string) {
@@ -78,4 +79,40 @@ export function splitLabel(optionLabel: string, inputValue: string) {
 	}
 
 	return results;
+}
+
+export function generateHighlightStyles(
+	inputValue?: string
+): Record<string, CSSProperties> {
+	const styles: Record<string, CSSProperties> = {};
+	if (!inputValue) return styles;
+
+	const characters = inputValue.toLowerCase().split('');
+
+	characters.forEach((_, index) => {
+		// When typing "abc"
+		// This generates things like [data-char="a"]+[data-char="b"]+[data-char="c"]
+		// to ensure we select consecutive elements
+		const baseSelector = characters
+			.slice(0, index + 1)
+			.map((char) => `[data-char="${char}"]`)
+			.join(' + ');
+
+		// This generates things like [data-char="a"]:has(+[data-char="b"]+[data-char="c"])
+		// to ensure we select earlier elements whose later siblings match
+		const hasSelector = characters
+			.slice(index + 1)
+			.map((char) => `+ [data-char="${char}"]`)
+			.join(' ');
+
+		const fullSelector = hasSelector
+			? `${baseSelector}:has(${hasSelector})`
+			: baseSelector;
+
+		styles[fullSelector] = {
+			fontWeight: 'bold',
+		};
+	});
+
+	return styles;
 }