Combobox was pretty slow… especially in dev mode. Partly because Downshift needs to render the list again on every interaction, but mostly because Box is slow and the filtering/string matching happened on every time.

This PR moves to native HTML elements with css props and uses CSS to match the string once at the top-level only.

[View preview](https://design-system.agriculture.gov.au/pr-preview/pr-1740)

## Checklist

**Preflight**

- [x] Prefix the PR title with the slug of the package or component - e.g. `accordion: Updated padding` or `docs: Updated header links`
- [x] Describe the changes clearly in the PR description
- [x] Read and check your code before tagging someone for review
- [x] Create a changeset file by running `yarn changeset`. [Learn more about change management](https://design-system.agriculture.gov.au/guides/change-management).

**Testing**

- [x] Manually test component in various modern browsers at various sizes (use [Browserstack](https://www.browserstack.com/))
- [x] Manually test component in various devices (phone, tablet, desktop)
- [x] Manually test component using a keyboard
- [ ] Manually test component using a screen reader
- [ ] Manually tested in dark mode
- [x] Component meets [Web Content Accessibility Guidelines (WCAG) 2.1 standards](https://www.w3.org/TR/WCAG21/)
- [x] Add any necessary unit tests (HTML validation, snapshots etc)
- [x] Run `yarn test` to ensure tests are passing. If required, run `yarn test -u` to update any generated snapshots.

**Documentation**

- [x] Create or update documentation on the website
- [ ] Create or update stories for Storybook
- [ ] Create or update stories for Playroom snippets

## Perf testing

Profiling Combox with 1170 items, here are Profiler stats of opening the combobox and then mousing over 2 items.

### main

```
phase, actualDuration, baseDuration: update 925.2000002861023 893.8999996185303
phase, actualDuration, baseDuration: nested-update 0 893.8999996185303
phase, actualDuration, baseDuration: update 887.3999996185303 877.9000005722046
phase, actualDuration, baseDuration: nested-update 0 877.9000005722046
phase, actualDuration, baseDuration: update 930.0000002384186 917.8999998569489
phase, actualDuration, baseDuration: nested-update 0 917.8999998569489
```

### this

```
phase, actualDuration, baseDuration: update 243.599999666214 187.59999752044678
phase, actualDuration, baseDuration: nested-update 0 187.59999752044678
phase, actualDuration, baseDuration: update 172.69999957084656 141.60000109672546
phase, actualDuration, baseDuration: nested-update 0 141.60000109672546
phase, actualDuration, baseDuration: update 154.50000023841858 127.90000009536743
phase, actualDuration, baseDuration: nested-update 0 127.90000009536743
```

Typing "afgha" and then deleting character by character:

### main

```
phase, actualDuration, baseDuration: update 1775.5000009536743 1744.1000044345856
phase, actualDuration, baseDuration: nested-update 0 1744.1000044345856
phase, actualDuration, baseDuration: update 4097.199998855591 4057.1999990940094
phase, actualDuration, baseDuration: nested-update 0 4057.1999990940094
phase, actualDuration, baseDuration: update 1842.9000000953674 1824.8999998569489
phase, actualDuration, baseDuration: nested-update 0 1824.8999998569489
phase, actualDuration, baseDuration: update 1001.8000009059906 990.4000012874603
phase, actualDuration, baseDuration: nested-update 0 990.4000012874603
phase, actualDuration, baseDuration: update 1209.5999991893768 1192.899997472763
phase, actualDuration, baseDuration: nested-update 0 1192.899997472763
phase, actualDuration, baseDuration: update 1392.2000007629395 1379.2000012397766
phase, actualDuration, baseDuration: nested-update 0 1379.2000012397766
phase, actualDuration, baseDuration: update 12.899999856948853 12.799999713897705
phase, actualDuration, baseDuration: nested-update 0 12.799999713897705
phase, actualDuration, baseDuration: update 18.09999990463257 17.90000009536743
phase, actualDuration, baseDuration: nested-update 0 17.90000009536743
phase, actualDuration, baseDuration: update 12.799999952316284 12.400000095367432
phase, actualDuration, baseDuration: nested-update 0 12.400000095367432
phase, actualDuration, baseDuration: update 10.199999809265137 10
phase, actualDuration, baseDuration: nested-update 0 10
phase, actualDuration, baseDuration: update 9.199999809265137 8.999999761581421
phase, actualDuration, baseDuration: nested-update 0 8.999999761581421
phase, actualDuration, baseDuration: update 8.899999856948853 8.599999904632568
phase, actualDuration, baseDuration: nested-update 0 8.599999904632568
phase, actualDuration, baseDuration: update 28.199999570846558 27.5
phase, actualDuration, baseDuration: nested-update 0 27.5
phase, actualDuration, baseDuration: update 45.200000047683716 44.20000076293945
phase, actualDuration, baseDuration: nested-update 0 44.20000076293945
phase, actualDuration, baseDuration: update 1609.1999990940094 1570.5000023841858
phase, actualDuration, baseDuration: nested-update 0 1570.5000023841858
phase, actualDuration, baseDuration: update 771.5000011920929 761.8000016212463
phase, actualDuration, baseDuration: nested-update 0 761.8000016212463
phase, actualDuration, baseDuration: update 1046.1999998092651 1021.9999992847443
phase, actualDuration, baseDuration: nested-update 0 1021.9999992847443
```

### this

```
phase, actualDuration, baseDuration: update 217.19999933242798 157.9999976158142
phase, actualDuration, baseDuration: nested-update 0 157.9999976158142
phase, actualDuration, baseDuration: update 203.49999976158142 173.09999918937683
phase, actualDuration, baseDuration: nested-update 0 173.09999918937683
phase, actualDuration, baseDuration: update 148.80000042915344 123.49999952316284
phase, actualDuration, baseDuration: nested-update 0 123.49999952316284
phase, actualDuration, baseDuration: update 148.2999997138977 125.30000019073486
phase, actualDuration, baseDuration: nested-update 0 125.30000019073486
phase, actualDuration, baseDuration: update 147.50000023841858 123.80000042915344
phase, actualDuration, baseDuration: nested-update 0 123.80000042915344
phase, actualDuration, baseDuration: update 150.40000009536743 126.3999993801117
phase, actualDuration, baseDuration: nested-update 0 126.3999993801117
phase, actualDuration, baseDuration: update 3 2.6999998092651367
phase, actualDuration, baseDuration: nested-update 0 2.6999998092651367
phase, actualDuration, baseDuration: update 6.699999809265137 6.399999380111694
phase, actualDuration, baseDuration: nested-update 0 6.399999380111694
phase, actualDuration, baseDuration: update 4.700000047683716 4.099999904632568
phase, actualDuration, baseDuration: nested-update 0 4.099999904632568
phase, actualDuration, baseDuration: update 3.200000047683716 3.0999999046325684
phase, actualDuration, baseDuration: nested-update 0 3.0999999046325684
phase, actualDuration, baseDuration: update 3.3000001907348633 3
phase, actualDuration, baseDuration: nested-update 0 3
phase, actualDuration, baseDuration: update 3 2.8999998569488525
phase, actualDuration, baseDuration: nested-update 0 2.8999998569488525
phase, actualDuration, baseDuration: update 5.400000095367432 4.200000286102295
phase, actualDuration, baseDuration: nested-update 0 4.200000286102295
phase, actualDuration, baseDuration: update 5 4.299999952316284
phase, actualDuration, baseDuration: nested-update 0 4.299999952316284
phase, actualDuration, baseDuration: update 195.99999952316284 142.7999987602234
phase, actualDuration, baseDuration: nested-update 0 142.7999987602234
phase, actualDuration, baseDuration: update 144.099999666214 121.40000104904175
phase, actualDuration, baseDuration: nested-update 0 121.40000104904175
phase, actualDuration, baseDuration: update 181.2999987602234 132.50000309944153
phase, actualDuration, baseDuration: nested-update 0 132.50000309944153
```