@ChrisLaneAU 2024-08-20
I think it's worth adding both a story and a docs entry to talk about and test/validate large datasets. Ideally a function that allows a specific number of items. I've just done some quick load testing with this, which produced 975 items:

```ts
() => {
	const [value, setValue] = React.useState(null);
	const [options] = React.useState(Array.from({ length: 5 }).flatMap(() => COUNTRY_OPTIONS));
	return (
		<Combobox
			label="Select country"
			hint="Start typing to see results"
			value={value}
			onChange={setValue}
			options={options}
		/>
	);
};
```