@stowball 2024-08-29
Hey, it also improves the accessibility in WHCM!

Before:

<img width="433" alt="image" src="https://github.com/user-attachments/assets/72d36de1-f96f-4a7e-aab2-b4b08ad540de">


After:

<img width="453" alt="image" src="https://github.com/user-attachments/assets/7830acaa-a388-472e-a482-20d4dec6bcad">